Implement \Nova\Database\Manager::clearConnections an fix the Tests after moving the Doctrine based DBAL from System to App/Packages
